Address "invalid escape sequence" warnings.
Before this change:

    $ make -B scowl.db
    ./util/patch.py < patch > scowl-tmp.txt
    rm -f scowl.db
    ./scowl create-db scowl.db < scowl-tmp.txt
    .../wordlist/libscowl/__init__.py:69: SyntaxWarning: invalid escape sequence '\+'
      wordPartRegex = re.compile(f'(\+?)({_wordRegex})([*@~!-]?)†?')
    .../wordlist/libscowl/__init__.py:656: SyntaxWarning: invalid escape sequence '\('
      m_ = re.fullmatch('\((.+)\)', w)

Python warns us twice about using the "\" character for an escape in a Python string, with an invalid following character.  In these two cases we want the "\" escape character to exist literally in the string, to be consumed by the regular expression engine.

This is just what the "raw string" Python feature is for ( https://docs.python.org/3/reference/lexical_analysis.html#escape-sequences -- keep literal backslashes in the string), adopted here.  Note that the (linked) documentation says "all unrecognized escape sequences are left in the string unchanged", so the existing code works with warnings, but also: "Changed in version 3.12: Unrecognized escape sequences produce a SyntaxWarning. In a future Python version they will be eventually a SyntaxError."  So while this is only a warning now, it will be more serious in the future.